refactor(jazz-next): share constraint signature helpers
