+++ jazz-next/src/JazzNext/Compiler/Runtime.hs
@@ -60,6 +60,8 @@ import JazzNext.Compiler.BuiltinCatalog
 import JazzNext.Compiler.CapabilityFacts
   ( concreteConstraintArgument,
     constraintFunctionArgumentTypes,
+    constraintSignatureTypeContainsClassParameter,
+    constraintSignatureTypesCompatible,
     qualifiedMethodKey,
     signaturePayloadConstraintType,
     substituteClassMethodSignature
@@ -1613,7 +1615,7 @@ runtimeMethodCandidateExactlyMatches classParameter methodSignature arguments (R
           suppliedGenericArgumentTypes = take suppliedArgumentCount genericArgumentTypes
           suppliedArgumentTypes = take suppliedArgumentCount argumentTypes
           targetArgumentPositions =
-            map (runtimeConstraintSignatureTypeContainsClassParameter classParameter) suppliedGenericArgumentTypes
+            map (constraintSignatureTypeContainsClassParameter classParameter) suppliedGenericArgumentTypes
        in suppliedArgumentCount <= length genericArgumentTypes
             && suppliedArgumentCount <= length argumentTypes
             && or targetArgumentPositions
@@ -1630,21 +1632,6 @@ runtimeMethodCandidateExactlyMatches classParameter methodSignature arguments (R
 runtimeExactCandidateArgumentMatches :: Bool -> ConstraintSignatureType -> RuntimeValue -> Bool
 runtimeExactCandidateArgumentMatches targetArgumentPosition signatureType runtimeValue =
   not targetArgumentPosition || runtimeValueExactlyMatchesConstraint signatureType runtimeValue
-
-runtimeConstraintSignatureTypeContainsClassParameter :: Text -> ConstraintSignatureType -> Bool
-runtimeConstraintSignatureTypeContainsClassParameter classParameter signatureType =
-  case signatureType of
-      ConstraintTypeApplication _ arguments ->
-        any (runtimeConstraintSignatureTypeContainsClassParameter classParameter) arguments
-      ConstraintTypeList innerType ->
-        runtimeConstraintSignatureTypeContainsClassParameter classParameter innerType
-      ConstraintTypeTuple elementTypes ->
-        any (runtimeConstraintSignatureTypeContainsClassParameter classParameter) elementTypes
-      ConstraintTypeFunction argumentType resultType ->
-        runtimeConstraintSignatureTypeContainsClassParameter classParameter argumentType
-          || runtimeConstraintSignatureTypeContainsClassParameter classParameter resultType
-      ConstraintTypeName typeName ->
-        identifierText typeName == classParameter
 
 runtimeValueExactlyMatchesConstraint :: ConstraintSignatureType -> RuntimeValue -> Bool
 runtimeValueExactlyMatchesConstraint signatureType runtimeValue =
@@ -1723,7 +1710,7 @@ runtimeValueMatchesConstraint :: ConstraintSignatureType -> RuntimeValue -> Bool
 runtimeValueMatchesConstraint signatureType runtimeValue =
   case runtimeValue of
     VTyped typeHint _ ->
-      runtimeConstraintTypesCompatible typeHint signatureType
+      constraintSignatureTypesCompatible typeHint signatureType
     _ ->
       case signatureType of
         ConstraintTypeName typeName ->
@@ -1734,7 +1721,7 @@ runtimeValueMatchesConstraint signatureType runtimeValue =
           case runtimeValue of
             VList elements maybeTypeHint ->
               case maybeTypeHint of
-                Just typeHint -> runtimeConstraintTypesCompatible typeHint signatureType
+                Just typeHint -> constraintSignatureTypesCompatible typeHint signatureType
                 Nothing -> all (runtimeValueMatchesConstraint elementType) elements
             _ -> False
         ConstraintTypeTuple elementTypes ->
@@ -1745,39 +1732,8 @@ runtimeValueMatchesConstraint signatureType runtimeValue =
             _ -> False
         ConstraintTypeFunction {} ->
           case runtimeValue of
-            VClosure _ _ _ (Just typeHint) _ -> runtimeConstraintTypesCompatible typeHint signatureType
+            VClosure _ _ _ (Just typeHint) _ -> constraintSignatureTypesCompatible typeHint signatureType
             _ -> isFunctionValue runtimeValue
-
-runtimeConstraintTypesCompatible :: ConstraintSignatureType -> ConstraintSignatureType -> Bool
-runtimeConstraintTypesCompatible leftType rightType =
-  case (leftType, rightType) of
-    (ConstraintTypeName leftName, ConstraintTypeName rightName) ->
-      runtimeConstraintNamesCompatible leftName rightName
-    (ConstraintTypeApplication leftName leftArguments, ConstraintTypeApplication rightName rightArguments)
-      | runtimeConstraintNamesCompatible leftName rightName,
-        length leftArguments == length rightArguments ->
-          and (zipWith runtimeConstraintTypesCompatible leftArguments rightArguments)
-    (ConstraintTypeList leftElementType, ConstraintTypeList rightElementType) ->
-      runtimeConstraintTypesCompatible leftElementType rightElementType
-    (ConstraintTypeTuple leftElementTypes, ConstraintTypeTuple rightElementTypes)
-      | length leftElementTypes == length rightElementTypes ->
-          and (zipWith runtimeConstraintTypesCompatible leftElementTypes rightElementTypes)
-    (ConstraintTypeFunction leftArgumentType leftResultType, ConstraintTypeFunction rightArgumentType rightResultType) ->
-      runtimeConstraintTypesCompatible leftArgumentType rightArgumentType
-        && runtimeConstraintTypesCompatible leftResultType rightResultType
-    _ -> False
-
-runtimeConstraintNamesCompatible :: Identifier -> Identifier -> Bool
-runtimeConstraintNamesCompatible leftName rightName =
-  normalizeRuntimeConstraintName (identifierText leftName)
-    == normalizeRuntimeConstraintName (identifierText rightName)
-
-normalizeRuntimeConstraintName :: Text -> Text
-normalizeRuntimeConstraintName typeName =
-  case typeName of
-    "Int" -> "Int64"
-    "Float" -> "Float64"
-    _ -> typeName
 
 runtimeValueMatchesTypeName :: Text -> RuntimeValue -> Bool
 runtimeValueMatchesTypeName typeName runtimeValue =
@@ -2830,7 +2786,7 @@ runtimeStructuralEquality :: RuntimeValue -> RuntimeValue -> Maybe Bool
 runtimeStructuralEquality leftValue rightValue =
   case (leftValue, rightValue) of
     (VTyped leftTypeHint leftInnerValue, VTyped rightTypeHint rightInnerValue)
-      | runtimeConstraintTypesCompatible leftTypeHint rightTypeHint ->
+      | constraintSignatureTypesCompatible leftTypeHint rightTypeHint ->
           runtimeStructuralEquality leftInnerValue rightInnerValue
       | otherwise ->
           Just False
